Getting rid of circular dependancy by moving ElectionConfig type to types.ts
