+++ lib/av_verifier.ts
@@ -4,11 +4,10 @@ import { randomKeyPair } from './av_client/generate_key_pair';
 import { signPayload } from './av_client/sign';
 import { decrypt } from './av_client/decrypt_vote';
 import { isValidPedersenCommitment } from './av_client/crypto/pedersen_commitment';
-import { CommitmentOpening, VerifierItem, BoardCommitmentOpeningItem, VoterCommitmentOpeningItem, BallotCryptogramItem } from './av_client/types';
+import { CommitmentOpening, VerifierItem, BoardCommitmentOpeningItem, VoterCommitmentOpeningItem, BallotCryptogramItem, ElectionConfig } from './av_client/types';
 
 import {
   fetchElectionConfig,
-  ElectionConfig,
   validateElectionConfig
 } from './av_client/election_config';
